Modularize and package Python scripts
