+++ script/git-hooks/pre-commit
@@ -8,18 +8,18 @@
 # To enable, symlink this file to '.git/hooks/pre-commit' like so:
 #   ln -s ../../script/git-hooks/pre-commit .git/hooks/pre-commit
 
-FORMATTER_PATH="./script/formatting/formatter.py"
+FORMATTER_COMMAND="python3 -m script.formatting.formatter"
 
 FILES=$(git diff --name-only HEAD --cached --diff-filter=d | grep '\.\(cpp\|h\)$')
 if [ -n "$FILES" ]; then
-  python3 script/validators/source_validator.py --files $FILES
+  python3 -m script.validators.source_validator --files $FILES
   RESULT=$?
   if [ $RESULT -ne 0 ]; then
     echo "***************************************"
     echo "******* Peloton Pre-Commit Hook *******"
     echo "***************************************"
-    echo "Use \"$FORMATTER_PATH -c -f\" to format all staged files."
-    echo "Or use \"$FORMATTER_PATH --no-verify\" to temporarily bypass the pre-commit hook."
+    echo "Use \"$FORMATTER_COMMAND -c -f\" to format all staged files."
+    echo "Or use \"$FORMATTER_COMMAND --no-verify\" to temporarily bypass the pre-commit hook."
     echo
     echo "Be aware that changed files have to be staged again!"
     echo "***************************************"
